chore: minor code optimization in `purge_jobs`

+++ datajoint/autopopulate.py
@@ -600,20 +600,24 @@ class AutoPopulate:
         """
         invalid_removed = 0
 
-        invalid_success = len(self.jobs & "status = 'success'") - len(self.target)
-        if invalid_success > 0:
-            for key, job_key in zip(*(self.jobs & "status = 'success'").fetch("KEY", "key")):
-                if not (self.target & job_key):
-                    (self.jobs & key).delete()
-                    invalid_removed += 1
+        success_query = self.jobs & {"table_name": self.target.table_name} & "status = 'success'"
+        if success_query:
+            invalid_success = len(success_query) - len(self.target)
+            if invalid_success > 0:
+                for key, job_key in zip(*success_query.fetch("KEY", "key")):
+                    if not (self.target & job_key):
+                        (self.jobs & key).delete()
+                        invalid_removed += 1
 
-        keys2do = self._jobs_to_do({}).fetch("KEY")
-        invalid_incomplete = len(self.jobs & "status != 'success'") - len(keys2do)
-        if invalid_incomplete > 0:
-            for key, job_key in zip(*(self.jobs & "status != 'success'").fetch("KEY", "key")):
-                if job_key not in keys2do:
-                    (self.jobs & key).delete()
-                    invalid_removed += 1
+        incomplete_query = self.jobs & {"table_name": self.target.table_name} & "status != 'success'"
+        if incomplete_query:
+            keys2do = self._jobs_to_do({}).fetch("KEY")
+            invalid_incomplete = len(incomplete_query) - len(keys2do)
+            if invalid_incomplete > 0:
+                for key, job_key in zip(*incomplete_query.fetch("KEY", "key")):
+                    if job_key not in keys2do:
+                        (self.jobs & key).delete()
+                        invalid_removed += 1
 
         logger.info(
             f"{invalid_removed} invalid jobs removed for `{to_camel_case(self.target.table_name)}`"
